# docs/openapi.yml
openapi: "3.0.3"
info:
  title: 云裳衣裳 — 时尚推荐平台 API
  description: |
    AI 驱动的智能时尚穿搭推荐平台。支持图片上传、CLIP 风格分析、
    天气感知推荐、知识库 RAG 对话和衣橱管理。
  version: "1.0.0"
  contact:
    name: Fashion Platform Team
    url: https://github.com/your-org/fashion-platform
servers:
  - url: http://localhost:5000
    description: 本地开发服务器
...
components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "错误描述"
      required: [success, error]
    UploadResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        username:
          type: string
        image_url:
          type: string
        local_path:
          type: string
        filename:
          type: string
        description:
          type: string
          description: 纯中文图片描述
        color:
          type: string
          description: 主色调（中文）
        style_tag:
          type: string
          description: 风格标签（中文）
        cloth_type:
          type: string
          description: 衣物类型（中文）
    RecommendItem:
      type: object
      properties:
        image_url:
          type: string
        similarity:
          type: number
        color:
          type: string
        description:
          type: string
        style_tag:
          type: string
        cloth_type:
          type: string
        weather_suggestion:
          type: string
        is_wardrobe:
          type: boolean
        wardrobe_item_id:
          type: integer
          nullable: true
        reason:
          type: string
          description: 推荐理由（结合用户画像 + 天气）
    RecommendResponse:
      type: object
      properties:
        success:
          type: boolean
        city:
          type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/RecommendItem"
        used_wardrobe:
          type: boolean
    WeatherInfo:
      type: object
      properties:
        success:
          type: boolean
        city:
          type: string
        temperature:
          type: number
        condition:
          type: string
          description: 天气状况（中文）
        suggestion:
          type: string
          description: 穿搭建议
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Flask-Login session cookie
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
...
  /recommendation/upload:
    post:
      tags: [Recommendation]
      summary: 上传衣物图片
      description: 上传一张衣物图片，自动提取颜色、风格标签和类型描述
      operationId: uploadCloth
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                cloth_image:
                  type: string
                  format: binary
                username:
                  type: string
                style:
                  type: string
              required: [cloth_image]
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          $ref: "#/components/responses/BadRequest"
  /recommendation/ai_recommend:
    post:
      tags: [Recommendation]
      summary: AI 穿搭推荐
      description: |
        基于 CLIP 余弦相似度的 AI 推荐引擎：
        1. 提取上传图片的 CLIP 特征向量
        2. 与本地图库 + 衣柜物品计算相似度
        3. 结合天气和用户画像生成推荐理由
      operationId: aiRecommend
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                local_path:
                  type: string
                  description: 上传图片的服务端绝对路径
                city:
                  type: string
                  description: 城市名称
                include_wardrobe:
                  type: boolean
                  default: false
                  description: 是否包含衣柜物品
                user_profile:
                  type: object
                  properties:
                    height:
                      type: number
                    weight:
                      type: number
                    age:
                      type: integer
                    gender:
                      type: string
                    body_shape:
                      type: string
                    skin_tone:
                      type: string
                    style_pref:
                      type: array
                      items:
                        type: string
              required: [local_path]
      responses:
        "200":
          description: 推荐结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendResponse"
  /recommendation/weather:
    post:
      tags: [Recommendation]
      summary: 获取城市天气
      operationId: getWeather
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                city:
                  type: string
              required: [city]
      responses:
        "200":
          description: 天气信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeatherInfo"

  # ==================== Wardrobe ====================
...